+++ back/brm-core/internal/app/app_interface.go
@@ -21,27 +21,27 @@ func New(coreRepo repo.CoreRepo, authCli grpcauth.AuthClient) App {
 }
 
 type CompanyApp interface {
-	GetCompany(ctx context.Context, id uint) (model.Company, error)
+	GetCompany(ctx context.Context, id uint64) (model.Company, error)
 	CreateCompanyAndOwner(ctx context.Context, company model.Company, owner model.Employee) (model.Company, model.Employee, error)
-	UpdateCompany(ctx context.Context, companyId uint, ownerId uint, upd model.UpdateCompany) (model.Company, error)
-	DeleteCompany(ctx context.Context, companyId uint, ownerId uint) error
+	UpdateCompany(ctx context.Context, companyId uint64, ownerId uint64, upd model.UpdateCompany) (model.Company, error)
+	DeleteCompany(ctx context.Context, companyId uint64, ownerId uint64) error
 }
 
 type EmployeeApp interface {
-	CreateEmployee(ctx context.Context, companyId uint, ownerId uint, employee model.Employee) (model.Employee, error)
-	UpdateEmployee(ctx context.Context, companyId uint, ownerId uint, employeeId uint, upd model.UpdateEmployee) (model.Employee, error)
-	DeleteEmployee(ctx context.Context, companyId uint, ownerId uint, employeeId uint) error
+	CreateEmployee(ctx context.Context, companyId uint64, ownerId uint64, employee model.Employee) (model.Employee, error)
+	UpdateEmployee(ctx context.Context, companyId uint64, ownerId uint64, employeeId uint64, upd model.UpdateEmployee) (model.Employee, error)
+	DeleteEmployee(ctx context.Context, companyId uint64, ownerId uint64, employeeId uint64) error
 
-	GetCompanyEmployees(ctx context.Context, companyId uint, employeeId uint, filter model.FilterEmployee) ([]model.Employee, error)
-	GetEmployeeByName(ctx context.Context, companyId uint, employeeId uint, ebn model.EmployeeByName) ([]model.Employee, error)
-	GetEmployeeById(ctx context.Context, companyId uint, employeeId uint, employeeIdToFind uint) (model.Employee, error)
+	GetCompanyEmployees(ctx context.Context, companyId uint64, employeeId uint64, filter model.FilterEmployee) ([]model.Employee, error)
+	GetEmployeeByName(ctx context.Context, companyId uint64, employeeId uint64, ebn model.EmployeeByName) ([]model.Employee, error)
+	GetEmployeeById(ctx context.Context, companyId uint64, employeeId uint64, employeeIdToFind uint64) (model.Employee, error)
 }
 
 type ContactApp interface {
-	CreateContact(ctx context.Context, ownerId uint, employeeId uint) (model.Contact, error)
-	UpdateContact(ctx context.Context, ownerId uint, contactId uint, upd model.UpdateContact) (model.Contact, error)
-	DeleteContact(ctx context.Context, ownerId uint, contactId uint) error
+	CreateContact(ctx context.Context, ownerId uint64, employeeId uint64) (model.Contact, error)
+	UpdateContact(ctx context.Context, ownerId uint64, contactId uint64, upd model.UpdateContact) (model.Contact, error)
+	DeleteContact(ctx context.Context, ownerId uint64, contactId uint64) error
 
-	GetContacts(ctx context.Context, ownerId uint, pagination model.GetContacts) ([]model.Contact, error)
-	GetContactById(ctx context.Context, ownerId uint, contactId uint) (model.Contact, error)
+	GetContacts(ctx context.Context, ownerId uint64, pagination model.GetContacts) ([]model.Contact, error)
+	GetContactById(ctx context.Context, ownerId uint64, contactId uint64) (model.Contact, error)
 }
