+++ back/transport-api/internal/model/core/employee.go
@@ -1,8 +1,8 @@
 package core
 
 type Employee struct {
-	Id           uint
-	CompanyId    uint
+	Id           uint64
+	CompanyId    uint64
 	FirstName    string
 	SecondName   string
 	Email        string
@@ -27,13 +27,13 @@ type FilterEmployee struct {
 	ByDepartment bool
 	Department   string
 
-	Limit  int
-	Offset int
+	Limit  uint
+	Offset uint
 }
 
 type EmployeeByName struct {
 	Pattern string
 
-	Limit  int
-	Offset int
+	Limit  uint
+	Offset uint
 }
