+++ back/transport-api/internal/ports/httpserver/core/employees/handlers.go
@@ -40,9 +40,9 @@ func AddEmployee(a app.App) gin.HandlerFunc {
 			return
 		}
 
-		employee, err := a.CreateEmployee(c, uint(companyId), uint(ownerId), core.Employee{
+		employee, err := a.CreateEmployee(c, companyId, ownerId, core.Employee{
 			Id:           0,
-			CompanyId:    uint(req.CompanyId),
+			CompanyId:    req.CompanyId,
 			FirstName:    req.FirstName,
 			SecondName:   req.SecondName,
 			Email:        req.Email,
@@ -122,14 +122,14 @@ func GetEmployeesList(a app.App) gin.HandlerFunc {
 				employeeId,
 				core.EmployeeByName{
 					Pattern: pattern,
-					Limit:   limit,
-					Offset:  offset,
+					Limit:   uint(limit),
+					Offset:  uint(offset),
 				},
 			)
 		} else {
 			var filter core.FilterEmployee
-			filter.Limit = limit
-			filter.Offset = offset
+			filter.Limit = uint(limit)
+			filter.Offset = uint(offset)
 			filter.JobTitle, filter.ByJobTitle = c.GetQuery("jobtitle")
 			filter.Department, filter.ByDepartment = c.GetQuery("department")
 			employees, err = a.GetCompanyEmployees(c,
@@ -190,7 +190,7 @@ func GetEmployee(a app.App) gin.HandlerFunc {
 		employee, err := a.GetEmployeeById(c,
 			companyId,
 			employeeId,
-			uint(id))
+			id)
 
 		switch {
 		case err == nil:
@@ -253,7 +253,7 @@ func UpdateEmployee(a app.App) gin.HandlerFunc {
 		employee, err := a.UpdateEmployee(c,
 			companyId,
 			ownerId,
-			uint(id),
+			id,
 			core.UpdateEmployee{
 				FirstName:  req.FirstName,
 				SecondName: req.SecondName,
@@ -314,7 +314,7 @@ func DeleteEmployee(a app.App) gin.HandlerFunc {
 		err = a.DeleteEmployee(c,
 			companyId,
 			ownerId,
-			uint(id))
+			id)
 		switch {
 		case err == nil:
 			c.JSON(http.StatusOK, employeeResponse{
